Render cards instead of tables
